fix build

+++ lib/grandpa/grandpa_integration_test.go
@@ -231,8 +231,8 @@ func TestGetPossibleSelectedAncestors_SameAncestor(t *testing.T) {
 	}
 
 	votes := gs.getVotes(prevote)
-	prevoted := make(map[common.Hash]uint32)
-	var blocks map[common.Hash]uint32
+	prevoted := make(map[common.Hash]uint64)
+	var blocks map[common.Hash]uint64
 
 	for _, curr := range leaves {
 		blocks, err = gs.getPossibleSelectedAncestors(votes, curr, prevoted, prevote, gs.state.threshold())
@@ -291,8 +291,8 @@ func TestGetPossibleSelectedAncestors_VaryingAncestor(t *testing.T) {
 	}
 
 	votes := gs.getVotes(prevote)
-	prevoted := make(map[common.Hash]uint32)
-	var blocks map[common.Hash]uint32
+	prevoted := make(map[common.Hash]uint64)
+	var blocks map[common.Hash]uint64
 
 	for _, curr := range leaves {
 		blocks, err = gs.getPossibleSelectedAncestors(votes, curr, prevoted, prevote, gs.state.threshold())
@@ -357,8 +357,8 @@ func TestGetPossibleSelectedAncestors_VaryingAncestor_MoreBranches(t *testing.T)
 	}
 
 	votes := gs.getVotes(prevote)
-	prevoted := make(map[common.Hash]uint32)
-	var blocks map[common.Hash]uint32
+	prevoted := make(map[common.Hash]uint64)
+	var blocks map[common.Hash]uint64
 
 	for _, curr := range leaves {
 		blocks, err = gs.getPossibleSelectedAncestors(votes, curr, prevoted, prevote, gs.state.threshold())
@@ -1195,7 +1195,7 @@ func TestGrandpaServiceCreateJustification_ShouldCountEquivocatoryVotes(t *testi
 			AuthorityID: v.Public().(*ed25519.PublicKey).AsBytes(),
 			Vote: types.GrandpaVote{
 				Hash:   bfcHash,
-				Number: uint32(bfcNumber),
+				Number: uint64(bfcNumber),
 			},
 		}
